# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: GustoPOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all staff users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a staff user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a staff user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a staff user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a staff user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /tax-rates/{category}:
    get:
      operationId: getTaxRateByCategory
      tags: [tax-rates]
      summary: Get tax rate by category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tax rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRate"
    post:
      operationId: updateTaxRate
      tags: [tax-rates]
      summary: Update tax rate
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaxRateBody"
      responses:
        "200":
          description: Updated tax rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateTaxRateResponse"
...
  /staff-shifts/{shiftId}:
    get:
      operationId: getStaffShifts
      tags: [staff-shifts]
      summary: Get staff shifts for a shift
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Staff shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffShift"
  /staff-shifts/clock-in:
    post:
      operationId: clockInStaff
      tags: [staff-shifts]
      summary: Clock in staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInBody"
      responses:
        "200":
          description: Clocked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockInResponse"
  /staff-shifts/clock-out:
    post:
      operationId: clockOutStaff
      tags: [staff-shifts]
      summary: Clock out staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutBody"
      responses:
        "200":
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockOutResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    StaffUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, head_bartender, bartender, server]
        language:
          type: string
          enum: [en, es]
        pin:
          type: string
        password:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - language
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, employee]
        language:
          type: string
          enum: [en, es]
        pin:
          type: string
        password:
          type: ["string", "null"]
      required:
        - firstName
        - role
        - language
        - pin
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, employee]
        language:
          type: string
          enum: [en, es]
        pin:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
...
    TaxRate:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        rate:
          type: number
        description:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - category
        - rate
        - isActive
...
    UpdateTaxRateBody:
      type: object
      properties:
        rate:
          type: number
        description:
          type: string
      required:
        - rate
    UpdateTaxRateResponse:
      type: object
      properties:
        success:
          type: boolean
        rate:
          $ref: "#/components/schemas/TaxRate"
      required:
        - success
        - rate
...
    StaffShift:
      type: object
      properties:
        id:
          type: string
        shiftId:
          type: string
        staffUserId:
          type: string
        staffName:
          type: string
        staffRole:
          type: string
        clockInAt:
          type: string
          format: date-time
        clockOutAt:
          type: string
          format: date-time
        breakStartAt:
          type: string
          format: date-time
        breakEndAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        hoursWorked:
          type: number
        breakMinutes:
          type: integer
      required:
        - id
        - shiftId
        - staffUserId
        - staffName
        - clockInAt
    ClockInBody:
      type: object
      properties:
        shiftId:
          type: string
        staffUserId:
          type: string
        notes:
          type: string
      required:
        - shiftId
        - staffUserId
    ClockInResponse:
      type: object
      properties:
        success:
          type: boolean
        shift:
          type: object
          properties:
            id:
              type: string
            staffName:
              type: string
            clockInAt:
              type: string
              format: date-time
            notes:
              type: string
          required:
            - id
            - staffName
            - clockInAt
      required:
        - success
        - shift
    ClockOutBody:
      type: object
      properties:
        staffShiftId:
          type: string
        notes:
          type: string
      required:
        - staffShiftId
    ClockOutResponse:
      type: object
      properties:
        success:
          type: boolean
        shift:
          type: object
          properties:
            staffName:
              type: string
            clockInAt:
              type: string
              format: date-time
            clockOutAt:
              type: string
              format: date-time
            hoursWorked:
              type: number
          required:
            - staffName
            - clockInAt
            - clockOutAt
            - hoursWorked
      required:
        - success
        - shift
...